+++ NEWS.md
@@ -1,7 +1,8 @@
-# pedquant 0.1.2.1
+# pedquant 0.1.3
 
-* fixed a bug in ed_nbs when the symbol provided is wrong.
+* Fixed a bug in ed_nbs when the symbol provided is wrong.
 * Added china district code in ed_code function.
+* Added a new function md_stock_adjust to adjust the OHLC prices for splits and dividend.
 
 # pedquant 0.1.2